requirements.txt
# Main
RPi.GPIO

# RFID
mfrc522

# Fingerprint
adafruit-circuitpython-fingerprint

# Camera
picamera

# 7 segment LED
raspberrypi-tm1637